+++ pkg/store/getter.go
@@ -1,9 +1,9 @@
 package store
 
 import (
-	"encoding/json"
 	"fmt"
 
+	"github.com/wklken/apisix-go/pkg/json"
 	"github.com/wklken/apisix-go/pkg/logger"
 	"github.com/wklken/apisix-go/pkg/resource"
 	"github.com/wklken/apisix-go/pkg/util"