+++ CHANGELOG.md
@@ -58,3 +58,4 @@
 - [x] attach the consumer 2024-04-29
 - [x] plugin consumer-restriction [consumer-restriction](https://apisix.apache.org/zh/docs/apisix/plugins/consumer-restriction/) 2024-04-29
 - [x] plugin http-logger [http-logger](https://apisix.apache.org/zh/docs/apisix/plugins/http-logger/) 2024-04-30
+- [x] plugin udp-logger [udp-logger](https://apisix.apache.org/zh/docs/apisix/plugins/udp-logger/) 2024-05-01